Need to find a different method to fix the stdint.h issue on Solaris systems which lack stdint.h and instead define uintX_t types in inttypes.h.  Undoing the change.
The lib builds, but the included programs don't include config.h which sets
the defines appropriately, so compilation of them fails.

Perhaps we need a freeradius-client.h.in which gets modified to fit the
system it's being installed on so that it doesn't include non-existant
header files on systems that don't have them.