Fix: initialise PRNG properly
When configuring the library through rc_config_init, the PRNG is not
initalised. On platforms with no /dev/urandom this might be a serious
security issue because passwords are not properly protected anymore.

So this commit introduces a new utility function rc_random which
ensures, that the PRNG is initialised before the first call to
random.

This commit fixes #82

Signed-off-by: Marcel Patzlaff <m.patzlaff@pilz.de>